requirements.txt
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
allure-pytest==2.13.2
